requirements.txt
beautifulsoup4==4.9.3
confusable-homoglyphs==3.2.0
Django==2.2.6
django-bootstrap3==14.2.0
django-crispy-forms==1.9.2
djangorestframework==3.12.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.4.1
